modify RootComponent から振る舞いを Container.tsx にうつした
